style(analyzer): class H sweep — unused code, stale hides, part-file setState, hot-path stat
- removed all 4 stale 'hide readingProgressProvider' clauses (symbol no
  longer exists in quran_provider; every screen already imports the
  real provider file directly)
- 16 unused 'final colors' locals (scoped scan), 3 write-only fields
  (_timeZonesReady + its call, _nisabThreshold), 2 dead private methods
  (_showPageInfo, _getLocationInfo)
- cache_store: promoted return needs no '!'; surah_reader: inner
  followAyah null check provably true (enclosed block gates it)
- part-file extensions wrap setState in a _patchUi in-class method
  (invalid_use_of_protected_member, zero runtime change)
- tilawat usableSize -> sync existsSync/lengthSync (hot per-ayah/scan
  probe; the lint's own rationale), cold callers stay async
- RadioListTile kept: RadioGroup postdates this repo's Dart 3.4 /
  Flutter 3.22 floor; NOTE added per owner rule
- notification_service/main: write-only tz flag removed (its test
  line in phase12 dropped — singleton identity is the real assertion)

+++ lib/features/quran/data/repository/quran_meta.dart
@@ -242,7 +242,7 @@ abstract final class QuranMeta {
   static int computeTotalAyahs() {
     var total = 0;
     for (final row in _surahTable) {
-      total += row[1] as int;
+      total += row[1];
     }
     return total;
   }
